molgenis-data-annotators/src/test/java/org/molgenis/data/annotation/entity/impl/ExacAnnotatorTest.java
package org.molgenis.data.annotation.entity.impl;

import static org.mockito.Mockito.mock;
import static org.molgenis.data.EntityMetaData.AttributeRole.ROLE_ID;
import static org.testng.Assert.assertEquals;
import static org.testng.Assert.assertFalse;
import static org.testng.Assert.assertTrue;

import java.util.Collections;
import java.util.Iterator;
import java.util.LinkedHashMap;
import java.util.Map;

import org.molgenis.data.DataService;
import org.molgenis.data.Entity;
import org.molgenis.data.annotation.AnnotationService;
import org.molgenis.data.annotation.RepositoryAnnotator;
import org.molgenis.data.annotation.resources.Resources;
import org.molgenis.data.annotation.resources.impl.ResourcesImpl;
import org.molgenis.data.annotator.websettings.ExacAnnotatorSettings;
import org.molgenis.data.support.DefaultEntityMetaData;
import org.molgenis.data.support.MapEntity;
import org.molgenis.data.vcf.VcfRepository;
import org.molgenis.util.ResourceUtils;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.context.annotation.Bean;
import org.springframework.test.context.ContextConfiguration;
import org.springframework.test.context.testng.AbstractTestNGSpringContextTests;
import org.testng.annotations.Test;

@ContextConfiguration(classes =
{ ExacAnnotatorTest.Config.class, ExacAnnotator.class })
public class ExacAnnotatorTest extends AbstractTestNGSpringContextTests
{

	@Autowired
	RepositoryAnnotator annotator;

	@Test
	public void testAnnotate()
	{
		DefaultEntityMetaData emdIn = new DefaultEntityMetaData("exac");
		emdIn.addAttribute(VcfRepository.CHROM, ROLE_ID);
		emdIn.addAttributeMetaData(VcfRepository.POS_META);
		emdIn.addAttributeMetaData(VcfRepository.REF_META);
		emdIn.addAttributeMetaData(VcfRepository.ALT_META);

		Entity inputEntity = new MapEntity(emdIn);
		inputEntity.set(VcfRepository.CHROM, "1");
		inputEntity.set(VcfRepository.POS, 13372);
		inputEntity.set(VcfRepository.REF, "G");
		inputEntity.set(VcfRepository.ALT, "C");

		Iterator<Entity> results = annotator.annotate(Collections.singletonList(inputEntity));
		assertTrue(results.hasNext());
		Entity resultEntity = results.next();
		assertFalse(results.hasNext());

		Map<String, Object> expectedMap = new LinkedHashMap<String, Object>();
		expectedMap.put(VcfRepository.CHROM, "1");
		expectedMap.put(VcfRepository.POS, 13372);
		expectedMap.put(VcfRepository.REF, "G");
		expectedMap.put(VcfRepository.ALT, "C");
		expectedMap.put(ExacAnnotator.EXAC_AF, "6.998e-05");
		Entity expectedEntity = new MapEntity(expectedMap);

		assertEquals(resultEntity.get(VcfRepository.CHROM), expectedEntity.get(VcfRepository.CHROM));
		assertEquals(resultEntity.get(VcfRepository.POS), expectedEntity.get(VcfRepository.POS));
		assertEquals(resultEntity.get(VcfRepository.REF), expectedEntity.get(VcfRepository.REF));
		assertEquals(resultEntity.get(VcfRepository.ALT), expectedEntity.get(VcfRepository.ALT));
		assertEquals(resultEntity.get(ExacAnnotator.EXAC_AF), expectedEntity.get(ExacAnnotator.EXAC_AF));
	}

	public static class Config
	{
		@Autowired
		private DataService dataService;

		@Bean
		public Entity exacAnnotatorSettings()
		{
			Entity settings = new MapEntity();
			settings.set(ExacAnnotatorSettings.Meta.EXAC_LOCATION,
					ResourceUtils.getFile(getClass(), "/exac/exac_test_set.vcf.gz").getPath());
			return settings;
		}

		@Bean
		public DataService dataService()
		{
			return mock(DataService.class);
		}

		@Bean
		public AnnotationService annotationService()
		{
			return mock(AnnotationService.class);
		}

		@Bean
		public Resources resources()
		{
			return new ResourcesImpl();
		}
	}
}
